backend auth required
 * switching to Flask blueprints
 * DRY up tests

+++ blog-site/backend/example_backend/comments/controller.py
@@ -2,30 +2,31 @@ import http
 from typing import List
 
 import flask
-from flask import jsonify, request
+from flask import Blueprint, jsonify, request
 
-from example_backend.app import app
 from example_backend.comments import dao
 from example_backend.exc.bad_request import BadRequest
 from example_backend.exc.not_found import NotFound
 from example_backend.posts import dao as posts_dao
-from example_backend.users.service import resolve_user
+from example_backend.users.service import auth_required
+
+blueprint = Blueprint("comment", __name__, url_prefix="/posts/<int:post_id>/comments")
 
 
-@app.route("/posts/<int:post_id>/comments", methods=("POST",))
-def create(post_id):
+@blueprint.route("", methods=("POST",))
+@auth_required(inject_user=True)
+def create(post_id, user):
     request_json = flask.request.json
 
     post = posts_dao.find(post_id)
     if not post:
         raise BadRequest("Could not find post")
     else:
-        user, _ = resolve_user()
         comment = dao.insert(post.id, request_json["content"], user.id)
         return comment_to_json(comment)
 
 
-@app.route("/posts/<int:post_id>/comments", methods=("GET",))
+@blueprint.route("", methods=("GET",))
 def get_comments(post_id):
     post = posts_dao.find(post_id)
     if post is None:
@@ -33,16 +34,19 @@ def get_comments(post_id):
     return comments_to_json(post.comments)
 
 
-@app.route("/posts/<int:post_id>/comments/<int:comment_id>", methods=("GET",))
+@blueprint.route("/<int:comment_id>", methods=("GET",))
 def get_comment(post_id, comment_id):
     post = posts_dao.find(post_id)
     if post is None:
         raise NotFound(f"No post found with id {post_id}")
     comment = dao.find(comment_id)
+    if comment is None:
+        raise NotFound(f"No comment found with id {comment_id}")
     return comment_to_json(comment)
 
 
-@app.route("/posts/<int:post_id>/comments/<int:comment_id>", methods=("PUT",))
+@blueprint.route("/<int:comment_id>", methods=("PUT",))
+@auth_required()
 def edit(post_id, comment_id):
     post = posts_dao.find(post_id)
     if post is None:
@@ -53,11 +57,15 @@ def edit(post_id, comment_id):
     return comment_to_json(updated)
 
 
-@app.route("/posts/<int:post_id>/comments/<int:comment_id>", methods=("DELETE",))
+@blueprint.route("/<int:comment_id>", methods=("DELETE",))
+@auth_required()
 def delete(post_id, comment_id):
     post = posts_dao.find(post_id)
     if post is None:
         raise NotFound(f"No post found with id {post_id}")
+    comment = dao.find(comment_id)
+    if comment is None:
+        raise NotFound(f"No comment found with id {comment_id}")
     dao.delete(comment_id)
     return "", http.HTTPStatus.NO_CONTENT
